Move classes to demorun specific subpackage
