+++ samples/notary-demo/src/main/kotlin/net/corda/notarydemo/BFTNotaryCordform.kt
@@ -14,7 +14,7 @@ import net.corda.node.services.transactions.minCorrectReplicas
 import net.corda.node.utilities.ServiceIdentityGenerator
 import net.corda.testing.ALICE
 import net.corda.testing.BOB
-import net.corda.testing.internal.*
+import net.corda.testing.internal.demorun.*
 
 fun main(args: Array<String>) = BFTNotaryCordform.runNodes()
