+++ samples/notary-demo/src/main/kotlin/net/corda/notarydemo/Clean.kt
@@ -1,6 +1,6 @@
 package net.corda.notarydemo
 
-import net.corda.testing.internal.clean
+import net.corda.testing.internal.demorun.clean
 
 fun main(args: Array<String>) {
     listOf(SingleNotaryCordform, RaftNotaryCordform, BFTNotaryCordform).forEach {
