+++ testing/node-driver/src/main/kotlin/net/corda/testing/internal/demorun/CordformUtils.kt
@@ -1,6 +1,6 @@
 @file:JvmName("CordformUtils")
 
-package net.corda.testing.internal
+package net.corda.testing.internal.demorun
 
 import net.corda.cordform.CordformDefinition
 import net.corda.cordform.CordformNode
